move functionality to StatsExporter
